+++ config.py
@@ -1,0 +1,7 @@
+class Config:
+    def __init__(self) -> None:
+        self.open_ai_key = ""
+
+    def get_open_ai_key(self):
+        return self.open_ai_key
+
